:art: 5/27 할당량
랜딩페이지 애니매이션, 다른유저 피드 클릭 페이지 구현, 파일 확장자 체크

+++ src/components/tempClothes/tempClothes.queries.ts
@@ -20,7 +20,7 @@ export const Q_FETCH_USER = gql`
   }
 `;
 export const Q_FETCH_FEED_TAGS = gql`
-  query fetchFeedTags($count: Float!) {
+  query fetchFeedTags($count: Int!) {
     fetchFeedTags(count: $count) {
       tagName
       feed {
